# Zserver.web/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "基于Gin + Vue + Element UI的前后端分离权限管理系统的接口文档\n添加qq群: 74520518 进入技术交流群 请备注，谢谢！",
    "title": "go-admin API",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://github.com/wenjianzhang/ZserverWeb/blob/master/LICENSE.md"
    },
    "version": "0.0.1"
  },
  "paths": {
    "/api/v1/config": {
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取JSON",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "配置"
        ],
        "summary": "修改配置",
        "parameters": [
          {
            "description": "body",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/models.SysConfig"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": -1, \"message\": \"添加失败\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/config/{configId}": {
      "get": {
        "security": [
          {
            "": []
          }
        ],
        "description": "获取JSON",
        "tags": [
          "配置"
        ],
        "summary": "获取配置",
        "parameters": [
          {
            "type": "integer",
            "description": "配置编码",
            "name": "configId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "$ref": "#/definitions/app.Response"
            }
          }
        }
      },
      "delete": {
        "description": "删除数据",
        "tags": [
          "配置"
        ],
        "summary": "删除配置",
        "parameters": [
          {
            "type": "integer",
            "description": "configId",
            "name": "configId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": -1, \"message\": \"删除失败\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/configKey/{configKey}": {
      "get": {
        "security": [
          {
            "": []
          }
        ],
        "description": "获取JSON",
        "tags": [
          "配置"
        ],
        "summary": "获取配置",
        "parameters": [
          {
            "type": "integer",
            "description": "configKey",
            "name": "configKey",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "$ref": "#/definitions/app.Response"
            }
          }
        }
      }
    },
    "/api/v1/configList": {
      "get": {
        "security": [
          {
            "": []
          }
        ],
        "description": "获取JSON",
        "tags": [
          "配置"
        ],
        "summary": "配置列表数据",
        "parameters": [
          {
            "type": "string",
            "description": "configKey",
            "name": "configKey",
            "in": "query"
          },
          {
            "type": "string",
            "description": "configName",
            "name": "configName",
            "in": "query"
          },
          {
            "type": "string",
            "description": "configType",
            "name": "configType",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "页条数",
            "name": "pageSize",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "页码",
            "name": "pageIndex",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "$ref": "#/definitions/app.Response"
            }
          }
        }
      }
    },
...
    "/api/v1/dept": {
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取JSON",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "部门"
        ],
        "summary": "修改部门",
        "parameters": [
          {
            "type": "integer",
            "description": "id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "body",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/models.Dept"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": -1, \"message\": \"添加失败\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取JSON",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "部门"
        ],
        "summary": "添加部门",
        "parameters": [
          {
            "description": "data",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/models.Dept"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": -1, \"message\": \"添加失败\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/dept/{deptId}": {
      "get": {
        "security": [
          {
            "": []
          }
        ],
        "description": "获取JSON",
        "tags": [
          "部门"
        ],
        "summary": "部门列表数据",
        "parameters": [
          {
            "type": "string",
            "description": "deptId",
            "name": "deptId",
            "in": "path"
          },
          {
            "type": "string",
            "description": "position",
            "name": "position",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "$ref": "#/definitions/app.Response"
            }
          }
        }
      }
    },
    "/api/v1/dept/{id}": {
      "delete": {
        "description": "删除数据",
        "tags": [
          "部门"
        ],
        "summary": "删除部门",
        "parameters": [
          {
            "type": "integer",
            "description": "id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": -1, \"message\": \"删除失败\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/deptList": {
      "get": {
        "security": [
          {
            "": []
          }
        ],
        "description": "分页列表",
        "tags": [
          "部门"
        ],
        "summary": "分页部门列表数据",
        "parameters": [
          {
            "type": "string",
            "description": "name",
            "name": "name",
            "in": "query"
          },
          {
            "type": "string",
            "description": "id",
            "name": "id",
            "in": "query"
          },
          {
            "type": "string",
            "description": "position",
            "name": "position",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "$ref": "#/definitions/app.Response"
            }
          }
        }
      }
    },
    "/api/v1/dict/data": {
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取JSON",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "字典数据"
        ],
        "summary": "修改字典数据",
        "parameters": [
          {
            "description": "body",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/models.DictType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": -1, \"message\": \"添加失败\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取JSON",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "配置"
        ],
        "summary": "添加配置",
        "parameters": [
          {
            "description": "data",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/models.SysConfig"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": -1, \"message\": \"添加失败\"}",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "/login": {
      "post": {
        "description": "获取token",
        "consumes": [
          "application/json"
        ],
        "summary": "登陆",
        "parameters": [
          {
            "description": "Add account",
            "name": "username",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "$ref": "#/definitions/models.Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"expire\": \"2019-08-07T12:45:48+08:00\", \"token\": \".eyJleHAiOjE1NjUxNTMxNDgsImlkIjoiYWRtaW4iLCJvcmlnX2lhdCI6MTU2NTE0OTU0OH0.-zvzHvbg0A\" }",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/logout": {
      "post": {
        "security": [
          {
            "": []
          }
        ],
        "description": "获取token",
        "consumes": [
          "application/json"
        ],
        "summary": "退出登录",
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"msg\": \"成功退出系统\" }",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "app.Response": {
      "type": "object",
      "properties": {
        "code": {
          "description": "代码",
          "type": "integer",
          "example": 200
        },
        "data": {
          "description": "数据集",
          "type": "object"
        },
        "msg": {
          "description": "消息",
          "type": "string"
        }
      }
    },
    "models.Dept": {
      "type": "object",
      "properties": {
        "children": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/models.Dept"
          }
        },
        "createBy": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "dataScope": {
          "type": "string"
        },
        "deletedAt": {
          "type": "string"
        },
        "deptId": {
          "description": "部门编码",
          "type": "integer"
        },
        "deptName": {
          "description": "部门名称",
          "type": "string"
        },
        "deptPath": {
          "type": "string"
        },
        "email": {
          "description": "邮箱",
          "type": "string"
        },
        "leader": {
          "description": "负责人",
          "type": "string"
        },
        "params": {
          "type": "string"
        },
        "parentId": {
          "description": "上级部门",
          "type": "integer"
        },
        "phone": {
          "description": "手机",
          "type": "string"
        },
        "sort": {
          "description": "排序",
          "type": "integer"
        },
        "status": {
          "description": "状态",
          "type": "string"
        },
        "updateBy": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "models.DictType": {
      "type": "object",
      "properties": {
        "createBy": {
          "description": "创建者",
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "dataScope": {
          "type": "string"
        },
        "deletedAt": {
          "type": "string"
        },
        "dictId": {
          "type": "integer"
        },
        "dictName": {
          "description": "字典名称",
          "type": "string"
        },
        "dictType": {
          "description": "字典类型",
          "type": "string"
        },
        "params": {
          "type": "string"
        },
        "remark": {
          "description": "备注",
          "type": "string"
        },
        "status": {
          "description": "状态",
          "type": "string"
        },
        "updateBy": {
          "description": "更新者",
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
    "models.Login": {
      "type": "object",
      "required": [
        "code",
        "password",
        "username",
        "uuid"
      ],
      "properties": {
        "code": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "uuid": {
          "type": "string"
        }
      }
    },
...
    "models.SysConfig": {
      "type": "object",
      "properties": {
        "configId": {
          "description": "编码",
          "type": "integer"
        },
        "configKey": {
          "description": "参数键名",
          "type": "string"
        },
        "configName": {
          "description": "参数名称",
          "type": "string"
        },
        "configType": {
          "description": "是否系统内置",
          "type": "string"
        },
        "configValue": {
          "description": "参数键值",
          "type": "string"
        },
        "createBy": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "dataScope": {
          "type": "string"
        },
        "deletedAt": {
          "type": "string"
        },
        "params": {
          "type": "string"
        },
        "remark": {
          "description": "备注",
          "type": "string"
        },
        "updateBy": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}